http_head_options change

+++ reportbook/httpscripts/delete_assessment.php
@@ -2,12 +2,11 @@
 /**                    httpscripts/delete_assessment.php
  */
 
-require_once('common.php');
+require_once('../../scripts/http_head_options.php');
 
-if(isset($_GET{'eid'})){$eid=$_GET{'eid'};}
-elseif(isset($_POST{'eid'})){$eid=$_POST{'eid'};}
-else{print "Failed"; exit;}
+if(!isset($xmlid)){print "Failed"; exit;}
 
+	$eid=$xmlid;
    	$d_midcid=mysql_query("DELETE midcid, eidmid FROM midcid JOIN eidmid ON
 	   	eidmid.mark_id=midcid.mark_id WHERE eidmid.assessment_id='$eid'");
    	$d_eidmid=mysql_query("DELETE FROM eidmid WHERE assessment_id='$eid'");
@@ -16,7 +15,7 @@ else{print "Failed"; exit;}
 
 $returnXML=fetchAssessmentDefinition($eid);
 $rootName='AssessmentDefinition';
-require_once('commonreturn.php');
+require_once('../../scripts/http_end_options.php');
 exit;
 ?>
